+++ src/core/channels.ts
@@ -1,11 +1,11 @@
 'use strict'
 import { Neovim } from '@chemzqm/neovim'
-import { Disposable } from '../util/protocol'
 import { URI } from 'vscode-uri'
 import events from '../events'
 import BufferChannel from '../model/outputChannel'
 import { TextDocumentContentProvider } from '../provider'
 import { OutputChannel } from '../types'
+import { Disposable } from '../util/protocol'
 
 class Channels {
   private outputChannels: Map<string, BufferChannel> = new Map()