+++ history.md
@@ -1,6 +1,7 @@
 # 2023-01-29
 
 - Exclude `source` actions when request code actions with range.
+- Any character can be used for channel name.
 
 # 2023-01-26